Consolidate duplicate MagicCard interface definitions
- Create shared types package @mtg-middle-school/shared-types
- Move all type definitions to central location
- Update all imports to use shared types package
- Remove duplicate interfaces from API and web apps
- Maintain backward compatibility with re-export files
- Add workspace dependencies for both API and web apps

Types consolidated:
- MagicCard (with perfectMatch field from web version)
- CardSearchResult, DeckEntry, DeckValidationResult, SearchParams
- PartialMagicCard for CSV parsing
