+++ apps/api/package.json
@@ -15,7 +15,8 @@
     "hono": "^4.0.0",
     "graphql": "^16.8.0",
     "graphql-yoga": "^5.0.0",
-    "@hono/node-server": "^1.0.0"
+    "@hono/node-server": "^1.0.0",
+    "@mtg-middle-school/shared-types": "workspace:*"
   },
   "devDependencies": {
     "@types/node": "^20.0.0",
