+++ apps/web/package.json
@@ -22,7 +22,8 @@
     "isbot": "^4.1.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-i18next": "^15.5.3"
+    "react-i18next": "^15.5.3",
+    "@mtg-middle-school/shared-types": "workspace:*"
   },
   "devDependencies": {
     "@remix-run/dev": "^2.0.0",
