+++ packages/shared-types/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "@mtg-middle-school/shared-types",
+  "version": "1.0.0",
+  "description": "Shared TypeScript interfaces for MTG Middle School applications",
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build": "tsc",
+    "typecheck": "tsc --noEmit"
+  },
+  "devDependencies": {
+    "typescript": "^5.0.0"
+  },
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js",
+      "require": "./dist/index.js"
+    }
+  }
+}
